created tests to cover archive url when it's set on config and on rendered links
